#OSIS-6 Move code which don't respect rule and remove cyclic dependecy
