+++ assistant/forms.py
@@ -28,6 +28,8 @@ from django.core.exceptions import ValidationError
 from django.forms import ModelForm, Textarea, ModelChoiceField
 from django.forms.models import inlineformset_factory
 from django.utils.translation import ugettext as _
+
+import base.models.entity
 from base.models import academic_year, entity, entity_version
 from base.models.enums import entity_type
 from assistant import models as mdl
@@ -86,7 +88,7 @@ class EntityChoiceField(ModelChoiceField):
 
 def get_field_qs(field, **kwargs):
     if field.name == 'entity':
-        return EntityChoiceField(queryset=entity_version.find_versions_from_entites(
+        return EntityChoiceField(queryset=base.models.entity.find_versions_from_entites(
             entity.search(entity_type=entity_type.SECTOR) |
             entity.search(entity_type=entity_type.FACULTY) |
             entity.search(entity_type=entity_type.SCHOOL) |
@@ -313,7 +315,7 @@ class ReviewerForm(ModelForm):
     entities = \
         entity.search(entity_type=entity_type.INSTITUTE) | entity.search(entity_type=entity_type.FACULTY) | \
         entity.search(entity_type=entity_type.SECTOR)
-    entity = EntityChoiceField(required=True, queryset=entity_version.find_versions_from_entites(entities, None))
+    entity = EntityChoiceField(required=True, queryset=base.models.entity.find_versions_from_entites(entities, None))
 
     class Meta:
         model = mdl.reviewer.Reviewer
