+++ base/models/entity.py
@@ -26,9 +26,11 @@
 from django.contrib import admin
 from django.core.exceptions import ObjectDoesNotExist
 from django.db import models
+from django.db.models import Case, When, Q, F
 from django.utils import timezone
 
 from base.models import entity_version
+from base.models.enums import entity_type
 
 
 class EntityAdmin(admin.ModelAdmin):
@@ -110,3 +112,15 @@ def _find_descendants(entity, date=None, with_entities=True):
                                  entity_vers.find_descendants(date=date)}
     finally:
         return entities_descendants
+
+
+def find_versions_from_entites(entities, date):
+    if date is None:
+        date = timezone.now()
+    order_list = [entity_type.SECTOR, entity_type.FACULTY, entity_type.SCHOOL, entity_type.INSTITUTE, entity_type.POLE]
+    preserved = Case(*[When(entity_type=pk, then=pos) for pos, pk in enumerate(order_list)])
+    return Entity.objects.filter(pk__in=entities).\
+        filter(Q(entityversion__end_date__gte=date) | Q(entityversion__end_date__isnull=True),
+               entityversion__start_date__lte=date).\
+        annotate(acronym=F('entityversion__acronym')).annotate(title=F('entityversion__title')).\
+        annotate(entity_type=F('entityversion__entity_type')).order_by(preserved)
